refactor: phase 5 - extract ai service and remove gemini logic from app
